requirements.txt
coverage==7.4.4
Django==5.0.4
djangorestframework==3.15.1
psycopg2==2.9.9
python-dotenv==1.0.1
